CHERI PURECAP RISCV KERNEL
This is an experimental code and work-in-progress.
The code is written with kernel codying style guideline in mind.
But as it work-in-progress, it may contain some odds things :)

Written by:
Dmitry Kasatkin <dmitry.kasatkin@huawei.com>
Kui Wang <wang.kui1@huawei.com>

+++ arch/riscv/include/asm/stacktrace.h
@@ -7,12 +7,12 @@
 #include <asm/ptrace.h>
 
 struct stackframe {
-	unsigned long fp;
-	unsigned long ra;
+	uintptr_t fp;
+	uintptr_t ra;
 };
 
 extern void notrace walk_stackframe(struct task_struct *task, struct pt_regs *regs,
-				    bool (*fn)(void *, unsigned long), void *arg);
+				    bool (*fn)(void *, uintptr_t), void *arg);
 extern void dump_backtrace(struct pt_regs *regs, struct task_struct *task,
 			   const char *loglvl);
