+++ drivers/tty/pty.c
@@ -461,7 +461,7 @@ static void pty_remove(struct tty_driver *driver, struct tty_struct *tty)
 }
 
 static int pty_bsd_ioctl(struct tty_struct *tty,
-			 unsigned int cmd, unsigned long arg)
+			 unsigned int cmd, uintptr_t arg)
 {
 	switch (cmd) {
 	case TIOCSPTLCK: /* Set PT Lock (disallow slave open) */
@@ -650,7 +650,7 @@ err:
 }
 
 static int pty_unix98_ioctl(struct tty_struct *tty,
-			    unsigned int cmd, unsigned long arg)
+			    unsigned int cmd, uintptr_t arg)
 {
 	switch (cmd) {
 	case TIOCSPTLCK: /* Set PT Lock (disallow slave open) */
