+++ drivers/tty/tty_jobctrl.c
@@ -565,7 +565,7 @@ err:
  * if not then tty == real_tty.
  */
 long tty_jobctrl_ioctl(struct tty_struct *tty, struct tty_struct *real_tty,
-		       struct file *file, unsigned int cmd, unsigned long arg)
+		       struct file *file, unsigned int cmd, uintptr_t arg)
 {
 	void __user *p = (void __user *)arg;
