+++ fs/ext4/mballoc.c
@@ -444,8 +444,8 @@ static inline u64 ext4_get_discard_pa_seq_sum(void)
 static inline void *mb_correct_addr_and_bit(int *bit, void *addr)
 {
 #if BITS_PER_LONG == 64
-	*bit += ((unsigned long) addr & 7UL) << 3;
-	addr = (void *) ((unsigned long) addr & ~7UL);
+	*bit += ((uintptr_t) addr & 7UL) << 3;
+	addr = (void *) ((uintptr_t) addr & ~7UL);
 #elif BITS_PER_LONG == 32
 	*bit += ((unsigned long) addr & 3UL) << 3;
 	addr = (void *) ((unsigned long) addr & ~3UL);
