+++ fs/pipe.c
@@ -603,7 +603,7 @@ out:
 	return ret;
 }
 
-static long pipe_ioctl(struct file *filp, unsigned int cmd, unsigned long arg)
+static long pipe_ioctl(struct file *filp, unsigned int cmd, uintptr_t arg)
 {
 	struct pipe_inode_info *pipe = filp->private_data;
 	int count, head, tail, mask;
