+++ include/linux/file.h
@@ -50,12 +50,12 @@ extern struct file *fget(unsigned int fd);
 extern struct file *fget_many(unsigned int fd, unsigned int refs);
 extern struct file *fget_raw(unsigned int fd);
 extern struct file *fget_task(struct task_struct *task, unsigned int fd);
-extern unsigned long __fdget(unsigned int fd);
-extern unsigned long __fdget_raw(unsigned int fd);
-extern unsigned long __fdget_pos(unsigned int fd);
+extern uintptr_t __fdget(unsigned int fd);
+extern uintptr_t __fdget_raw(unsigned int fd);
+extern uintptr_t __fdget_pos(unsigned int fd);
 extern void __f_unlock_pos(struct file *);
 
-static inline struct fd __to_fd(unsigned long v)
+static inline struct fd __to_fd(uintptr_t v)
 {
 	return (struct fd){(struct file *)(v & ~3),v & 3};
 }
