+++ include/linux/genhd.h
@@ -282,7 +282,7 @@ int __invalidate_device(struct block_device *bdev, bool kill_dirty);
 void set_capacity(struct gendisk *disk, sector_t size);
 
 /* for drivers/char/raw.c: */
-int blkdev_ioctl(struct block_device *, fmode_t, unsigned, unsigned long);
+int blkdev_ioctl(struct block_device *, fmode_t, unsigned, uintptr_t);
 long compat_blkdev_ioctl(struct file *, unsigned, unsigned long);
 
 #ifdef CONFIG_BLOCK_HOLDER_DEPRECATED
