+++ include/linux/initrd.h
@@ -15,7 +15,7 @@ extern unsigned long rd_size;
 extern int initrd_below_start_ok;
 
 /* free_initrd_mem always gets called with the next two as arguments.. */
-extern unsigned long initrd_start, initrd_end;
+extern uintptr_t initrd_start, initrd_end;
 extern void free_initrd_mem(unsigned long, unsigned long);
 
 #ifdef CONFIG_BLK_DEV_INITRD
