+++ include/linux/key.h
@@ -149,12 +149,12 @@ typedef struct __key_reference_with_attributes *key_ref_t;
 static inline key_ref_t make_key_ref(const struct key *key,
 				     bool possession)
 {
-	return (key_ref_t) ((unsigned long) key | possession);
+	return (key_ref_t) ((uintptr_t) key | possession);
 }
 
 static inline struct key *key_ref_to_ptr(const key_ref_t key_ref)
 {
-	return (struct key *) ((unsigned long) key_ref & ~1UL);
+	return (struct key *) ((uintptr_t) key_ref & ~1UL);
 }
 
 static inline bool is_key_possessed(const key_ref_t key_ref)
