+++ include/linux/net.h
@@ -154,7 +154,7 @@ struct proto_ops {
 	__poll_t	(*poll)	     (struct file *file, struct socket *sock,
 				      struct poll_table_struct *wait);
 	int		(*ioctl)     (struct socket *sock, unsigned int cmd,
-				      unsigned long arg);
+				      uintptr_t arg);
 #ifdef CONFIG_COMPAT
 	int	 	(*compat_ioctl) (struct socket *sock, unsigned int cmd,
 				      unsigned long arg);
