+++ include/linux/percpu-refcount.h
@@ -107,7 +107,7 @@ struct percpu_ref {
 	 * The low bit of the pointer indicates whether the ref is in percpu
 	 * mode; if set, then get/put will manipulate the atomic_t.
 	 */
-	unsigned long		percpu_count_ptr;
+	uintptr_t		percpu_count_ptr;
 
 	/*
 	 * 'percpu_ref' is often embedded into user structure, and only
@@ -158,7 +158,7 @@ static inline void percpu_ref_kill(struct percpu_ref *ref)
 static inline bool __ref_is_percpu(struct percpu_ref *ref,
 					  unsigned long __percpu **percpu_countp)
 {
-	unsigned long percpu_ptr;
+	uintptr_t percpu_ptr;
 
 	/*
 	 * The value of @ref->percpu_count_ptr is tested for
